Update brukertaus.m
Updated brukertaus.m with more robust code for detecting the tau values of 4P and 6P HYSCORE experiments. It assumes (as the original did) that the 4P HYSCORE tau value (and first tau for 6P HYSCORE) is specified as d1 in the Bruker DSC file, and that the second tau for 6P HYSCORE is d2. The main change is replacing string searches with MATLAB regexp calls that simplify the extraction of values from text.